camera/requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
defusedxml==0.7.1
dlib @ file:///D:/nova-sigap/camera/dlib-19.24.1-cp311-cp311-win_amd64.whl#sha256=6f1a5ee167975d7952b28e0ce4495f1d9a77644761cf5720fb66d7c6188ae496
dnspython==2.7.0
dotenv==0.9.9
face-recognition==1.3.0
face_recognition_models==0.3.0
filelock==3.18.0
Flask==3.1.0
fonttools==4.57.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
lap==0.5.12
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
narwhals==1.34.1
networkx==3.4.2
numpy==1.24.4
opencv-python==4.11.0.86
packaging==24.2
paho-mqtt==2.1.0
pandas==2.2.3
pillow==11.2.1
protobuf==5.29.4
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==19.0.1
pydeck==0.9.1
pymongo==4.12.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
streamlit==1.44.1
supervision==0.25.1
sympy==1.13.1
tenacity==9.1.2
toml==0.10.2
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.107
ultralytics-thop==2.0.14
urllib3==2.4.0
watchdog==6.0.0
Werkzeug==3.1.3
